Increase size of row accumulator
I was accumulating `uint64_t` into `int`, which may not fit.

+++ sqlgrep/sqlgrep.cpp
@@ -187,7 +187,7 @@ auto find_and_display_matches(string_view to_find, int const maximum_results_per
     sql.set_logger(new database_query_logger);  // `new` is required by SOCI
     auto all_columns = get_all_string_columns(sql);
     cout << "Searching " << all_columns.size() << " columns for '" << to_find << "'...\n";
-    uint64_t const total_rows = accumulate(begin(all_columns), end(all_columns), 0, [](int acc, column_details const & b) { return acc + b.number_of_rows; });
+    uint64_t const total_rows = accumulate(begin(all_columns), end(all_columns), 0, [](uint64_t acc, column_details const & b) { return acc + b.number_of_rows; });
     write_verbose("Total number of rows to search: "s + to_string(total_rows) + ".");
     display_all_matches(sql, all_columns, to_find, maximum_results_per_column, total_rows);
 }
